packages/bh-sentinel-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bh-sentinel-core"
version = "0.1.0"
description = "Clinical decision support: pattern-based safety signal detection for behavioral health. Not a diagnostic tool."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "BH Healthcare", email = "oss@bh-healthcare.github.io" }
]
keywords = [
    "nlp",
    "clinical",
    "safety",
    "behavioral-health",
    "healthcare",
    "hipaa",
    "signal-detection",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0.0,<3",    # Request/response model validation
    "pyyaml>=6.0,<7",        # Pattern library YAML loading
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<9",
    "pytest-asyncio>=0.21,<1",   # Async Pipeline test support
    "ruff>=0.1.0,<1",
    "mypy>=1.0.0,<2",
]

[project.urls]
Homepage = "https://github.com/bh-healthcare/bh-sentinel"
Documentation = "https://github.com/bh-healthcare/bh-sentinel#readme"
Repository = "https://github.com/bh-healthcare/bh-sentinel"
Issues = "https://github.com/bh-healthcare/bh-sentinel/issues"
Changelog = "https://github.com/bh-healthcare/bh-sentinel/blob/main/CHANGELOG.md"
Author = "https://tanmayakumar.com"

[project.scripts]
bh-sentinel = "bh_sentinel.cli.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bh_sentinel"]
only-include = ["src/bh_sentinel/core", "src/bh_sentinel/cli"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/CHANGELOG.md",
    "/LICENSE",
]

[tool.hatch.build]
artifacts = [
    "*.json",
    "*.yaml",
    "py.typed",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
